refactor(layer-manager): Move remove added layers to the dialog, use eventBus for interaction with draw
